Refactor auth into server
