--- internal/controller/rest/api/v0/doc.go
@@ -1,4 +1,0 @@
-// Package api contains all API handlers and defines the ErrorHandler
-// interface, alongside an embeddable ErrorHandler implementation providing
-// common helper methods.
-package v0
