--- internal/controller/rest/api/v0/users/doc.go
@@ -1,2 +1,0 @@
-// Package users exposes a api for users endpoints.
-package users
