+++ internal/domain/user/doc.go
@@ -1,0 +1,6 @@
+// Package user specifies:
+//   - User-specific domain types;
+//   - A domain interface and corresponding implementation, for injection
+//     into inbound methods;
+//   - A outbound interface describing a store of user data.
+package user
