+++ internal/domain/user/errors.go
@@ -1,0 +1,179 @@
+package user
+
+import (
+	"errors"
+	"fmt"
+	"github.com/google/uuid"
+	"strings"
+)
+
+// FieldType identifies the type of field implicated in an error.
+type FieldType int
+
+const (
+	IDFieldType FieldType = iota + 1
+	UsernameFieldType
+	EmailFieldType
+	PasswordFieldType
+	URLFieldType
+)
+
+var fieldNames = [5]string{"id", "username", "email", "password", "image URL"}
+
+func (f FieldType) String() string {
+	return fieldNames[f-1]
+}
+
+// AuthError is a wrapper for authentication errors, which may include errors
+// that would otherwise be considered validation errors. This reinforces the
+// security convention that an end user should not receive the specifics of why
+// an authentication request failed.
+type AuthError struct {
+	Cause error
+}
+
+func (e *AuthError) Error() string {
+	return "unauthorized"
+}
+
+func (e *AuthError) Unwrap() error {
+	return e.Cause
+}
+
+// NotFoundError should be returned by a [Repository] when the specified user
+// does not exist.
+type NotFoundError struct {
+	IDField FieldType
+	ID      string
+}
+
+func (e *NotFoundError) Error() string {
+	return fmt.Sprintf("user with %s %q not found", e.IDField, e.ID)
+}
+
+func NewNotFoundByIDError(id uuid.UUID) error {
+	return &NotFoundError{
+		IDField: IDFieldType,
+		ID:      id.String(),
+	}
+}
+
+func NewNotFoundByEmailError(email EmailAddress) error {
+	return &NotFoundError{
+		IDField: EmailFieldType,
+		ID:      email.String(),
+	}
+}
+
+// ValidationErrors is a collection of [ValidationError], allowing us to report
+// on the validation of several inputs rather than returning the first validation
+// error encountered.
+type ValidationErrors []*ValidationError
+
+// PushValidationError adds a [ValidationError] to the collection. If the error
+// is not a ValidationError (including nil errors), it is returned as-is. The
+// supports the following pattern for successive validations:
+//
+//		var validationErrs ValidationErrors
+//		field1, err := parseField1()
+//		if err := validationErrs.PushValidationError(err); err != nil {
+//	     	return nil, err // not a ValidationError
+//		}
+//	 	field2, err := parseField2()
+//	 	...
+func (e *ValidationErrors) PushValidationError(err error) error {
+	var validationErr *ValidationError
+	if errors.As(err, &validationErr) {
+		*e = append(*e, validationErr)
+		return nil
+	}
+	return err
+}
+
+// Any returns true if the collection contains any errors, and false otherwise.
+func (e ValidationErrors) Any() bool {
+	return len(e) > 0
+}
+
+func (e ValidationErrors) Error() string {
+	var builder strings.Builder
+	builder.WriteString("validation errors:\n")
+	for _, err := range e {
+		builder.WriteString(fmt.Sprintf("\t- %s\n", err.Error()))
+	}
+	return builder.String()
+}
+
+// ValidationError represents a single error encountered when validation a field
+// of the given [FieldType].
+type ValidationError struct {
+	FieldType FieldType
+	Message   string
+}
+
+func (e *ValidationError) Error() string {
+	return fmt.Sprintf("%s: %s", e.FieldType, e.Message)
+}
+
+func NewUsernameTooShortError() error {
+	return &ValidationError{
+		FieldType: UsernameFieldType,
+		Message:   fmt.Sprintf("must be at least %d characters long", UsernameMinLen),
+	}
+}
+
+func NewUsernameTooLongError() error {
+	return &ValidationError{
+		FieldType: UsernameFieldType,
+		Message:   fmt.Sprintf("must be at most %d characters long", UsernameMaxLen),
+	}
+}
+
+func NewUsernameFormatError() error {
+	return &ValidationError{
+		FieldType: UsernameFieldType,
+		Message:   fmt.Sprintf("must match %q", UsernamePattern),
+	}
+}
+
+func NewDuplicateUsernameError(username Username) error {
+	return &ValidationError{
+		FieldType: UsernameFieldType,
+		Message:   fmt.Sprintf("%q is already registered", username),
+	}
+}
+
+func NewEmailAddressFormatError(candidate string) error {
+	return &ValidationError{
+		FieldType: EmailFieldType,
+		Message:   fmt.Sprintf("%q is not a valid email address", candidate),
+	}
+}
+
+func NewDuplicateEmailError(email EmailAddress) error {
+	return &ValidationError{
+		FieldType: EmailFieldType,
+		Message:   fmt.Sprintf("%q is already registered", email),
+	}
+}
+
+func NewPasswordTooShortError() error {
+	return &ValidationError{
+		FieldType: PasswordFieldType,
+		Message:   fmt.Sprintf("must be at least %d bytes long", PasswordMinLen),
+	}
+}
+
+func NewPasswordTooLongError() error {
+	return &ValidationError{
+		FieldType: PasswordFieldType,
+		Message:   fmt.Sprintf("must be at most %d bytes long", PasswordMaxLen),
+	}
+}
+
+func NewInvalidURLError() error {
+	return &ValidationError{
+		FieldType: URLFieldType,
+		Message:   "must be a valid URL",
+	}
+}
