+++ internal/inbound/rest/api/v0/users/doc.go
@@ -1,0 +1,2 @@
+// Package users exposes an api for users endpoints.
+package users
