+++ internal/outbound/sqlite/doc.go
@@ -1,0 +1,3 @@
+// Package sqlite provides an API to an SQLite3 database that conforms to domain
+// outbound interfaces.
+package sqlite
