+++ internal/outbound/sqlite/queries/queries.sql
