--- internal/repository/sqlite/doc.go
@@ -1,3 +1,0 @@
-// Package sqlite provides an API to an SQLite3 database that conforms to domain
-// repository interfaces.
-package sqlite
