--- internal/service/user/doc.go
@@ -1,6 +1,0 @@
-// Package user specifies:
-//   - User-specific domain types;
-//   - A service interface and corresponding implementation, for injection
-//     into controller methods;
-//   - A repository interface describing a store of user data.
-package user
