+++ pkg/option/option.go
@@ -1,6 +1,7 @@
 package option
 
 import (
+	"encoding/json"
 	"errors"
 )
 
@@ -8,6 +9,20 @@ import (
 type Option[T any] struct {
 	some  bool
 	value T
+}
+
+func (o Option[T]) UnmarshalJSON(bytes []byte) error {
+	if len(bytes) == 0 {
+		return nil
+	}
+
+	if err := json.Unmarshal(bytes, &o.value); err != nil {
+		return err
+	}
+
+	o.some = true
+
+	return nil
 }
 
 // Some returns an Option[T] populated with the given type.
@@ -46,4 +61,22 @@ func (o Option[T]) Value() (T, error) {
 // if the Option is None.
 func (o Option[T]) ValueOrZero() T {
 	return o.value
+}
+
+// Map transforms an Option[T] into an Option[U] by applying the given transform function.
+// None[T] is returned as None[U].
+//
+// # Errors
+//   - Any error returned by the transform function.
+func Map[T any, U any](opt Option[T], transform func(T) (U, error)) (Option[U], error) {
+	if !opt.Some() {
+		return None[U](), nil
+	}
+
+	u, err := transform(opt.ValueOrZero())
+	if err != nil {
+		return None[U](), err
+	}
+
+	return Some(u), nil
 }
