+++ tasks/Makefile.docker
@@ -65,4 +65,14 @@ docker/volume/ls: docker/volume/create
 	--mount type=volume,src=${REALWORLD_VOLUME_NAME},dst=/tmp/vol \
 	--user ${REALWORLD_USER}:${REALWORLD_USER} \
 	busybox \
-	ls -la /tmp
+	ls -la /tmp/vol
+
+docker/volume/manage: docker/volume/create
+	docker run \
+	--interactive \
+	--tty \
+	--rm \
+	--workdir /tmp/vol \
+	--mount type=volume,src=${REALWORLD_VOLUME_NAME},dst=/tmp/vol \
+	busybox \
+	sh
